Customize trust CI template
Modifies the trust CI templates with project specific configuration.

All "TODO" items are replaced with "DONE".

Initially, this has us building and testing on the stable channel on a
variety of platforms.

Included are Android and iOS, which we won't support initially. However,
it'll be good to keep these building, so we don't introduce gratuitious
incompatibilities.